checkmate/backend/actionsAddItem/requirements.txt
boto3==1.35.14
PyJWT==2.9.0
